requirements.txt
python-dotenv
langchain
langchain-azure-ai
langgraph
